Reconciles this branch with 7ee6b57: main's PR #3 already reworked
NotificationService (kIsWeb/TargetPlatform/Firebase.apps) and several
repositories, so those parts of this branch's fix commit are dropped in favour
of main's version. Kept from this branch: the two nested-class compile errors
main still has (SupabaseConfig.StorageBuckets, ProductFormView FileRef), the
SearchController/Flutter name clash, the missing Pdf/Export/Image service
registrations, the offline customer sync handler, the push-token guard, the
unused dart:io import in NotificationService (web builds) and cold-start
getInitialMessage handling, plus README/docs.

# Conflicts:
#	lib/features/auth/controllers/session_controller.dart
#	lib/services/notification_service.dart

+++ lib/services/image_service.dart
@@ -45,7 +45,7 @@ class ImageService {
         );
       }
       final Uint8List encoded =
-          img.encodeJpg(resized, quality: quality.clamp(1, 100));
+          img.encodeJpg(resized, quality: quality.clamp(1, 100).toInt());
       final File out = File(
         '${Directory.systemTemp.path}/img_${IdGenerator.uuid()}.jpg',
       );
